requirements.txt
# ─── Core data models ─────────────────────────────────────────────────────────
pydantic>=2.0
python-dotenv>=1.0.0

# ─── Kafka ────────────────────────────────────────────────────────────────────
kafka-python>=2.0.2

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
sqlalchemy>=2.0

# ─── Graph analytics ──────────────────────────────────────────────────────────
networkx>=3.0
python-louvain>=0.16
pandas>=2.0.0
neo4j>=5.0

# ─── ML / Anomaly detection ───────────────────────────────────────────────────
scikit-learn>=1.4.0
pyod>=1.1.0

# ─── API ──────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# ─── Utils ────────────────────────────────────────────────────────────────────
tenacity>=8.0.0
structlog>=24.0.0
joblib>=1.3.0

# ─── Caching ──────────────────────────────────────────────────────────────────
redis>=5.0.0

# ─── Observability ────────────────────────────────────────────────────────────
prometheus-fastapi-instrumentator>=6.1.0

# ─── Dev / Testing ────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
hypothesis>=6.0.0
locust>=2.20.0

# NOTE: apache-airflow is NOT listed here.
# It is installed exclusively inside the Airflow Docker container.
# Installing it locally conflicts with many of the packages above.
